[mlir] Add resource to transform dialect for external libraries
This resource enables referencing an external file and have it loaded
into the context while still being able to reference it in the IR.
Referencing into the external libraries follows the convention of
resource key followed by nested symbols inside.

This doesn't yet add the attribute type to reference resource. It would
effectively look the same as SymbolRefAttr but consist of two parts
where the former is referencing the resource and the remainder the
nested references. At that point the builder could be changed to only
retain referenced external files.