[mlir] Walk nested tables in symbol dce
The previous positioning was effectively that a symbol is dead if it
cannot be addressed from top level. I think that is too strong a
requirement: one can have operations that one cannot delete/DCE that
refers to symbols which one could delete. This resulted in symbol-dce
deleting symbols that are still referenced and the resulting IR being
invalid.

This instead all the symbols of top level operations of non-symbol table
ops additionally, as those are either dead and DCE would have handled,
or alive and we cannot just delete symbols referenced internally. E.g.,
this treats non-symbol table regioned ops more conservatively.